初始化

+++ src/ht/model/Model.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type Model struct {
-	ID         uint      `gorm:"primarykey"`
+	ID         int64     `gorm:"primarykey"`
 	CreateTime time.Time `gorm:"autoCreateTime"`
 	UpdateTime time.Time `gorm:"autoUpdateTime"`
 }
